pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "dst_airlines"
version = "0.1.0"
dependencies = [
    "requests",
    "python-dotenv",
    "cryptography"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]